Skip scenario tests if HA router will not be active
Due to related bug, it may happen that HA router will not become active
on any of the L3 agents even in 10 minutes.
I know it's just ugly workaround, but to unblock our gate lets skip test
if that issue will happen, instead of failing whole job.
We know that issue and we are already working on fix for that.

Related-Bug: #1923633
Change-Id: I185a0d1031af9489731baa0f132762b61eb9c64f