Add test for mac learning
This scenario test checks that, even with no DHCP and no port security,
packets are not flooded to every port on the network.

To see this work in OVN, port security and DHCP must be disabled on a
network. Before having those capabilities, packets would be flooded
to all servers. Now, the OVN switch can learn the MAC of a port and
store it on the FDB table. The adoption of this feature was completed
in [0].

[0] https://review.opendev.org/c/openstack/neutron/+/763567

Change-Id: I0313ac13f8e8922849f6388d2d1379817230294b