Remove allowed_address_pair API tests
The same tests are also available in
tempest.api.network.test_allowed_address_pair module. Usually in
such case, tests are removed from the tempest repo and kept in the
tempest plugin repo but in this case we discussed that with the core
members of the QA team and we agreed that it will be better to remove
those tests from neutron-tempest-plugin and keep them in the tempest
repo as they are part of the Interop tests.

Change-Id: Ia1756ddbea665dea639de8ef40602034c07dadbb