Fix required extensions for the subnet's service type API tests
In [1] new tests for subnet service types were added but by mistake
'service-types' was set as API extension required for those tests.
In fact this API feature in the subnet is provided by the
'subnet-service-types' extension and this should be required for that
tests.

This patch also adds "subnet-service-types" extension to the Tempest
conf in jobs for all branches up to stable/queens becuase this API
extension was available in Neutron since queens for sure so it's safe to
add it there.

[1] https://review.opendev.org/c/openstack/neutron-tempest-plugin/+/787718

Related-Bug: #1933115
Change-Id: I2b9b472a07455f5e67fc4adcc3b264b4ad8a9dbf