Use an older version of neutron-tempest-plugin for stein jobs
Since [0], stein uses an older version of tempest to fix jsonschema
issue observed in newer releases. This tag does not have
utils.is_network_feature_enabled used in test_metadata scenario.

As stein is in Extended Maintenance, use an older version for n-t-p to
fix gates. Also filter out port forwarding test as it is unstable in
this pinned version.

Closes-Bug: #1946748

[0] https://review.opendev.org/c/openstack/devstack/+/812092

Change-Id: I724da80bc440cd5d5b947d0c0a08cbead95628e9