test_list_agent: pop 'alive' from agent dict
As the test class lists agents in resource setup it can happen that the
selected agent from the list is not 'alive', but when test
test_list_agent lists again the agents, it is already up and
'alive'=True.
To avoid such failure, do skip 'alive' field of the comparison.

Change-Id: I0ed012ef59550e853a30e2994a1ab53083d8640c
Closes-Bug: #1953480