Exclude test test_floatingip_port_details from the Linuxbridge job
During the last PTG [1] we decided that due to lack of maintainers for
the Linuxbridge backend we will probably mark it as "unsupported" (or
something like that, nameing isn't decided yet) thus we will not
guarantee that all tests with such backend will always run properly and
all will be in good shape.
As a first "victim" of that decision, lets skip test
neutron_tempest_plugin.scenario.test_floatingip.FloatingIPPortDetailsTest.\
test_floatingip_port_details
which is failing from time to time in the linuxbridge scenario job.

Related-bug: #1799790

[1] https://etherpad.opendev.org/p/neutron-zed-ptg#L171

Change-Id: I0b481bc6b66fbe7f5e4dab92e9f60812bbaeaead