Bump swap to 2GB in linuxbridge/openvswitch jobs
Currently these jobs are running with 1GB swap
and fails sometimes with oom kill as node goes out
of memory.

Doubling the space to 2GB should help in avoiding
these random failures.

The issue seems to be triggered after api/scenario
jobs are combined[1]. So bumping only for the affected
jobs and also not bumping to 4GB/8GB because
in that case we may hide future actual issues.

[1] https://review.opendev.org/c/openstack/neutron-tempest-plugin/+/836912

Related-Bug: #1940243
Change-Id: Icded7e5a3b6484178ea2c9b145068788f23554e2