Update nested-virt testing for the 2023.1 cycle
Since the new main platform for testing is now Ubuntu Jammy, switch the
nested-virt jobs as well. In order to be able to keep the previous setup
for stable/zed and older branches, we add explicit nodeset overrides for
the affected jobs.

Cirros 0.5.1 seems to have issues booting on Jammy hosts, update to the
latest stable release 0.5.2.

Depends-on: https://review.opendev.org/c/openstack/devstack/+/860795
Change-Id: Ie835afa301e7fd17fb751b932adac74141e0e9a4