Using "Designate API" in addition to DNS query client check
As of now the only validation done in Neutron&Designate E2E scenario
test cases is based on DNS query client only.
As a result, we don't really know what's happens on the Designate side,
for example, we wouldn't expect to see the recordset on Backend if it
wasn't previously created on the Designate.
This patch improves the situation and it will definitely help on debugging.
It will use "Designate APIs" when "query client" throws an assertion.
For example, if "A" type recordset wasn't detected on backend (as expected
by test), "Designate APIs" will be used to provide more details on failure.

Change-Id: Ie2093e457e0b664f36204d61f3020ef84ef15a3e