[Secure RBAC] Add scope enforcement enabled job for master branch
This patch adds one new job

neutron-tempest-plugin-openvswitch-enforce-scope-new-defaults

which have enabled using scope enforcing and new default API policies
in Neutron and other projects (Glance, Nova) which supports that.

In next cycle (2023.2) we will want to make new policies and scope
enforcing to be enabled by default in Neutron but for now lets enable
it in just one job to make sure it's working and is ready to switch in
the future.

Change-Id: Ib51f39e83fbc4cfd2097998f957a76b0ae380caf