Remove note about migration from lib/neutron-legacy to lib/neutron
In [1] we finally got rid of the unfinished lib/neutron module and kept
only lib/neutron-legacy. It's renamed to lib/neutron now and it's the
only neutron related module in Devstack.
So this patch removes left over todo comments about things to do when
migration to the new lib/neutron module will be finished.

[1] https://review.opendev.org/c/openstack/devstack/+/865014

Change-Id: I73169ab4e660ddb18cb10acc783ec97a049665c9