Pin neutron-tempest-plugin for stable/wallaby jobs
As stable/wallaby is in EM phase, and currently broken
with the Tempest master. We are pinning tempest there (depends-on).

We are using the tempest 29.0.0 on stable/wallaby (depends-on) and
compatible tag for tempest 29.0.0 is neutron-tempest-plugin 1.8.0
(as neutron-tempest-plugin >= 1.9.0 need tempest>-29.2.0) 

- https://github.com/openstack/neutron-tempest-plugin/blob/1.9.0/requirements.txt#L14
- https://zuul.opendev.org/t/openstack/build/f644e4c4c2234ba981c56d91a5b11cbf

As stable/wallaby is in EM phase, we can remove the wallaby job
from master gate.

Depends-On: https://review.opendev.org/c/openstack/devstack/+/871782
Change-Id: Ide5789b40180d572459d47dbd474433644e4dcd7