[Stateless SG] Add IPv6 and DHCP related tests
When stateless SG are used RA packets may be blocked as there is no
stateful connections in SG so it will not be by default accepted
as related to the outgoing RS packets.
To confirm that this works fine, this patch adds new, dual stack cases
for the test "test_default_sec_grp_scenarios". This test is the same as
existing test_default_sec_grp_scenarios for stateful and stateless IPv4
scenarios but additionally checks at the end if IPv6 addresses were
configured properly in the guest VMs.

Change-Id: I1595e5c2072f43e9c8fd7cbfe6fa25815f1ffd92
Depends-On: Id1328d7cba418fa7c227ae9db4fe83c09fd06035