[Ussuri] Fix broken gate
Do not use ubuntu minimal image in ussuri scenario jobs.
The minimal image does not have the guestmount package
available, so use the server image instead.

Also had to add an override for designate-tempest-plugin
to address a similar problem that was fixed in the train
template in review:
https://review.opendev.org/c/openstack/neutron-tempest-plugin/+/817465

In addition, to fix the iptables_hybrid job, backport a
change to exclude a security group test that should not run,
test_established_tcp_session_after_re_attachinging_sg,
taken from review:
https://review.opendev.org/c/openstack/neutron-tempest-plugin/+/873708

Change-Id: Iac70ae2b2cb6fe97b96f4d93de19cf08ed144724