Fix the way how default SG for project if found in SG scenario test
Previously it was just getting list of SGs and assumed that first one in
the list is the "default" one. It's not always true so now there is
helper method which lists of SGs for project and looks for the one with
name "default" as this name is hardcoded in Neutron.

Change-Id: I608677547c9c8ae00af8821622cefa0955d692c6