Get nc client's command based on the nc on correct host
As nc client's command depends on the nc version installed
it should be checked on the host where it is going to be run.
Before this patch it was always checked on the host where tempest
was running and that could cause some issues e.g. when tempest
was run on RHEL 8/Centos 8 host and Cirros was used for guest VMs.

Change-Id: Ib0ce72a1b622d0f6fcda24f4b78f627092edf639