gate: bump ovn to the latest LTS release (22.03)
21.03 is no longer supported upstream. We should try to stick to LTS
releases of OVN if possible.

The patch locks the version as the latest git commit in branch-22.03
(ovs is also locked to the corresponding submodule version). This is
done to include a just merged fix for protocols used for ipv6 address
configuration (ra, na, mld*).

Once 22.03.3 is tagged in OVN repo, we can switch to it.

Depends-On: I83fc9250ae5b7c1686938a0dd25d66b40fc6c6aa
Depends-On: I2a633741d5e95f0e46a6b33198903f0f44d449b6
Change-Id: I6d7f6f4fbabc6273b170dfb963a8a37200e9fbc2