Don't run stable/xena jobs in check/gate queue anymore
Branch stable/xena was transitioned to EM phase already [1] so we don't
need to run its jobs in check and gate queues of neutron-tempest-plugin
repo anymore.
In stable/xena jobs in Neutron and stadium projects we will be running
neutron-tempest-plugin 2.3.0 version which was released when stable/xena
was going to EM phase.

[1] https://review.opendev.org/c/openstack/releases/+/878874

Depends-On: https://review.opendev.org/c/openstack/releases/+/881230
Change-Id: I45cc1a4529b35fff2118ce6278147e1564f74f6b