Revert "Revert "Update nested-virt testing for the 2023.1 cycle""
This reverts commit 3c30984a53005ed2d7a6a2d37f304bbd631be62d.

All the pending hypervisors are upgraded in
vexxhost-ca-ymq-1 and that fixes the nested-virt
issue.
There is currently mirror issue[1] which is being
investigated but since the vexxhost provider is
disabled[2] we can switch jobs to jammy.

re enablement of vexxhost provider once mirror
issue is resolved shouldn't impact our jobs.

[1] https://bugs.launchpad.net/neutron/+bug/2017992
[2] https://review.opendev.org/c/openstack/project-config/+/881810

Related-Bug: #1999249
Change-Id: I60b7d94da0774558b35794fde522fa82c0259422