[S-RBAC] Update DSCP marking rule create API test
In the test_qos_dscp_create_and_update API test, qos policy was made as
admin user (which is correct) but was also owned by admin project. And
later, to check if DSCP marking rule was created in that policy
properly, regular client is used instead.
The problem is that with new S-RBAC API policies, rules are visible to
owners of the policy, not to all users. And due to that this test is
failing with new S-RBAC policies enforced.

This patch fixes it by changing owner of the qos policy to the regular
client's project.

Related-Bug: #2018727
Change-Id: Iadf69c167cdda0017084e482a58116520a1ea80f