Don't check exit status when nc_client is spawned
In all cases where we are using nc_client, we rely on the messages
returned there in stdout and we don't really check exit_status codes.

It seems that e.g. ncat provided by RHEL 9 returns error_code=1 in case
when connection was refused. And that is expected thing e.g. in some
parts of the port_forwarding related tests.
To avoid unexpected failures due to ShellCommandFailed error we
shouldn't check exit_code there and always return messages from
nc_client's stdout to assert them in the test.

Closes: rhbz#2210233
Change-Id: I17cebaa34acb0dd6c60bd106eb70656faaebbb04