Add neutron-tempest-plugin-linuxbridge job to the experimental queue
It was agreed on the Neutron CI meeting [1] that we will remove
this job from check and gate queues and add it to the experimental
queue. It is also added to the periodic queue in Neutron with [2].

Linuxbridge agent is marked as experimental since at least one cycle
already and this scenario job is getting less and less stable now so
it's probably time to move it out from our gate.

[1] https://meetings.opendev.org/meetings/neutron_ci/2023/neutron_ci.2023-06-27-15.00.log.html#l-127
[2] https://review.opendev.org/c/openstack/neutron/+/887158

Change-Id: If4233a01575dfc9ffc84295f032ed19e32c38a7b