Add username in resource_setup
There is a condition when rhel is used as default image
in tempest, so default_image_is_advanced set as true
and validation.image_ssh_user is set but
advanced_image_ssh_user is not set.
test_fragmented_traffic_is_accepted was failing
because advanced_image_ssh_user was empty.

Change-Id: I7fa5b167ba27cc78e017765eca927a1dfb24825e