Skip HA router tests when not meant for OVN driver
Change is needed to check "l3-ha" extension with OVN router flavors.
This change allows to check if OVN backend or not, and add test skips
accordingly.

The tests used to run only against OVS driver, a subset of them was
found useful also with OVN, for instance this bug noticed by altered
tests:
https://bugzilla.redhat.com/show_bug.cgi?id=2263225

Change-Id: I7c2c2f23417b3b43a2788b3c4bca2f8b532b7974