Add a new client method: "get_loaded_network_extensions"
This method retrieves the extensions loaded by the Neutron API, instead
of reading the configured extensions in the tempest configuration.

This method can help during a extension implementation, as for example
in "subnet-external-network". This extension cannot be declared yet in
the neutron-tempest-plugin job definitions, because the extension has
not been implemented in Neutron. However, when the Neutron patch is
tested, the extension is missing in the configuration and the test
cannot correctly modify the API calls.

Related-Bug: #2051831
Change-Id: I3e6f570cd63e44df7ceb8f5013c2b4adebfe4286