Make neutron-tempest-plugin-designate-scenario non voting
neutron-tempest-plugin-designate-scenario is failing 100%
with the designate new RBAC
- https://review.opendev.org/c/openstack/neutron/+/926085

This job is non voting on neutron side but voting on designate
side. And merging the requirement change
(https://review.opendev.org/c/openstack/requirements/+/925464),
will break the designate gate.

Making this job non voting for everyone until we figure out the
root cause of this failure.

Related-Bug: #2078518
Change-Id: If0f521a74ce075b408b5495a0144f01e3bcbc818