Add test for extension "quota-check-limit-default"
This new test is checking that, by default, the quota engine will check
the resource usage before setting the quota. If the new quota is below
the resource usage, an exception is returned.

This patch is also adding the flag "force=True" in other tests setting
quotas for resources; these tests do not need the quota engine to
verify the current resource usage.

Related-Bug: #1953170
Change-Id: Ia193cd5a2aacc4243b5807eb7757b32e66f12365