Change the default image in the NDR tests
This patch changes the default reporitory (hub.docker.com) to quay.io,
in order to avoid the limitations implemented in the website, reducing
the download request.

This patch is also changing the default image used (Ubuntu 22.04) to
Ubuntu 20.04 because the package "quagga-bgpd" is present only in the
older version.

NOTE: it will be needed to upgrade these tests in order to use a newer
OS release; it will be probably needed to use FRR instead of Quagga.

This patch has the same approach as [1]

[1]https://review.opendev.org/c/openstack/neutron-tempest-plugin/+/935743

Related-Bug: #2089140
Change-Id: I3b314fc818606d60e2acd00d0e8284f740a687e0