Add scenario tests for the Vlan QinQ feature
This patch adds 2 new scenario tests to test connectivity between VMs
using vlan network with enabled QinQ and with vlan configured inside
VMs.
This is very similar to the "vlan_transparent" option which was already
tested since some time. The only difference is in the ethtype of the
outer vlan packets send through such network. Unfortunately we can't
test ethtype of the outer vlan in the neutron-tempest-plugin tests as
that would require access to the host where VMs runs and test packets on
the physical NICs.
So those new tests are actually the same as the existing tests for the
network with "vlan_transparent" enabled to at least make sure that QinQ
functionality works from the end user point of view.

This patch also enables those new tests for the ML2/OVN job as this is
currently only backend with implemented support for QinQ.

Depends-On: https://review.opendev.org/c/openstack/neutron/+/937633
Related-Bug: #1915151
Change-Id: Icc2b175cf4cfe5c4ca73bb7ccc5a9089a27798c9