[FWaaS] Update icmp_reachability_test for OVN backend
FWaaS is using stateless ACL rules in OVN when OVN driver is used. That
means that traffic for both directions needs to be allowed explicitely
always.
This patch adjusts
fwaas.scenario.test_fwaas_v2.TestFWaaS_v2.test_icmp_reachability_scenarios
test so that in the 3rd step it allows ICMP traffic in both directions
and test is passing.

This patch also adds new config option `driver` in the `fwaas` group.
This option can be used to tell Tempest what FWaaS driver is used in the
deployment so that appropriate test path can be choosen.

Change-Id: I4f238300069df4777245bb64f83824ea36296cc7