Skip multi GW router update if IP in use
Skip the test ``test_create_router_update_external_gateways`` execution
if the router gateway update returns a ``IpAddressAlreadyAllocated``
exception. That could happen if the removed gateway port IP, between
the removal and the re-assignation to the same router, is allocated by
other test, as reported in the bug. The probabilities of this error are
low but not zero.

Closes-Bug: #2117383
Signed-off-by: Rodolfo Alonso Hernandez <ralonsoh@redhat.com>
Change-Id: I8038c5775cf5f1c20a4349ddc472415b65945219