Fix FloatingIpTestCasesAdmin.test_two_vms_fips
the test is overly optimistic, it takes a first az from the az list,
and first hypervisor from hypervisor list, and tries to boot server
with this combination.
But this is really not a given IRL that this hypervisor actually
belongs to this az, and the test will fail with NoValidHost.

Walk over detailed az list, and pick up a suitable nova-compute
host and node that is up and enabled.

Closes-Bug: #2136147
Change-Id: Ia436ead6537a9900235f93077ce63f02a3f3a385
Signed-off-by: Pavlo Shchelokovskyy <shchelokovskyy@gmail.com>